# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HSD Auth Platform API
  description: |
    HSD Auth Platform - Authentication as a Service API
    
    Bu API, HSD projelerinde kimlik doğrulama işlemleri için kullanılır.
    
    ## Authentication
    API istekleri için JWT Bearer token veya API Key kullanılabilir.
    
    ## Rate Limiting
    - Varsayılan: 100 istek/dakika
    - Rate limit aşıldığında 429 Too Many Requests döner
  version: 1.0.0
  contact:
    name: HSD Auth Support
    email: support@hsdcore.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.auth.hsdcore.com
    description: Production server
  - url: https://api.staging.auth.hsdcore.com
    description: Staging server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Kullanıcı girişi
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Giriş başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Token yenileme
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token yenilendi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    Unauthorized:
      description: Kimlik doğrulama gerekli
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate limit aşıldı
      headers:
        Retry-After:
          schema:
            type: integer
          description: Saniye cinsinden bekleme süresi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        status:
          type: string
          enum: [active, disabled, pending]
        profile:
          $ref: '#/components/schemas/Profile'
        roles:
          type: array
          items:
            type: string
        mfaEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
        metadata:
          type: object
...